Moved updateDB, and removed redundant ucmetadata.go
